% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{full_get_params}
\alias{full_get_params}
\title{Get LMM parameters for data with no missing values}
\usage{
full_get_params(Y, X, REML = TRUE)
}
\arguments{
\item{Y}{Matrix with individuals as rows and measurements as columns
Missing measurements must be NA.}

\item{X}{Matrix with \code{n} rows and \code{c} columns for \code{n}
individuals and \code{c} covariates.}

\item{REML}{Boolean for returning REML or ML estimates}
}
\value{
A list with ve in slot \code{ve}, vg in slot \code{vg}, 
        estimated effect sizes and uncorrelated standard errors
        in slot \code{coef}.
}
\description{
Get LMM parameters for data with no missing values
}
