% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{get_constants}
\alias{get_constants}
\title{Calculate constants independent of delta in MLE and REMLE likelihoods}
\usage{
get_constants(Y, X)
}
\arguments{
\item{Y}{Matrix with individuals as rows and measurements as columns
Missing measurements must be NA.}

\item{X}{Matrix with \code{n} rows and \code{c} columns for \code{n}
individuals and \code{c} covariates.}
}
\value{
A list of named constants. These constants are difficult to
        describe briefly. See Jew et al, 2020 for a more in-depth 
        discussion of these constants. \code{ns} is the total number of 
        measured responses. \code{ni} is the number of individuals.
        \code{ntc} is the number of measurement contexts multiplied
        by the number of covariates. \code{g.sizes} is a vector of 
        the unique number of measurements across all individuals.
        \code{g.ind.sizes} is a vector of the number of individuals
        that have a certain number of measurements. \code{B} is a dummy
        matrix that will store the inverse of XHX (beta hat covariance).
        \code{D} and \code{E} are vector representations of the upper 
        triangle of symmetric matrices that are used to calculate \code{B}.
        \code{sxy} is a vector of the sum of responses multiplied by each
        covariate within each context. \code{gxsy} is a matrix where each
        row represents a group and column represents the sum of the 
        covariate in each context multiplied by the sum of responses for
        each individual. \code{sy2} is the sum of squared responses.
        \code{syi2g} provides the sum of squared responses within each
        group of measurement numbers.
}
\description{
Calculate constants independent of delta in MLE and REMLE likelihoods
}
