% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{mc_approx}
\alias{mc_approx}
\title{Ultra fast approximation for multi-context LMM}
\usage{
mc_approx(Y, X)
}
\arguments{
\item{Y}{Matrix with individuals as rows and measurements as columns
Missing measurements must be NA}

\item{X}{Matrix with \code{n} rows and \code{c} columns for \code{n}
individuals and \code{c} covariates.}
}
\value{
List of approximate MLE parameters in slots \code{ve},
        and \code{vg}.
}
\description{
Ultra fast approximation for multi-context LMM
}
