% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{mc_remle}
\alias{mc_remle}
\title{Ultra-fast REMLE for multi-context LMM}
\usage{
mc_remle(Y, X)
}
\arguments{
\item{Y}{Matrix with individuals as rows and measurements as columns
Missing measurements must be NA}

\item{X}{Matrix with \code{n} rows and \code{c} columns for \code{n}
individuals and \code{c} covariates.}
}
\value{
List of REMLE parameters in slots \code{delta}, \code{ve},
        and \code{vg}.
}
\description{
Ultra-fast REMLE for multi-context LMM
}
