% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{mle_nll}
\alias{mle_nll}
\title{Calculate MLE negative log likelihood given detla}
\usage{
mle_nll(d, const)
}
\arguments{
\item{d}{Value of delta to evaluate}

\item{const}{A named list of constants used to calculate likelihood as
described in documentation of \code{get_constants}. These
variables are loaded into the environment with \code{list2env}.}
}
\value{
Negative log likelihood of MLE model given delta (float)
}
\description{
Calculate MLE negative log likelihood given detla
}
