% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mvrnorm}
\alias{mvrnorm}
\title{file MASS/R/mvrnorm.R
copyright (C) 1994-2015 W. N. Venables and B. D. Ripley}
\usage{
mvrnorm(n, mu, Sigma)
}
\description{
This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 or 3 of the License
 (at your option).
}
\details{
This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 A copy of the GNU General Public License is available at
 http://www.r-project.org/Licenses/

 Slightly edited code to remove some input checks.
 
 @param n Number of samples
 @param mu Mean vector
 @param Sigma Covariance matrix
 
 @return MVN sample in \code{n} by \code{length(mu)} matrix
}
