% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{remle_get_params_meta}
\alias{remle_get_params_meta}
\title{Calculate optimal variance component REMLE estimates given delta using Lin
and Sullivan approach or standard for Meta-Tissue}
\usage{
remle_get_params_meta(d, const, newRE = TRUE)
}
\arguments{
\item{d}{Value of delta to evaluate}

\item{const}{A named list of constants used to calculate parameters as
described in documentation of \code{get_constants}. These
variables are loaded into the environment with \code{list2env}.}
}
\value{
A list with ve in slot \code{ve}, vg in slot \code{vg}, 
        estimated effect sizes and uncorrelated standard errors
        in slot \code{coef}.
}
\description{
Calculate optimal variance component REMLE estimates given delta using Lin
and Sullivan approach or standard for Meta-Tissue
}
