% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate test data for mcLMM}
\usage{
simulate_data(ni, context.weights = NULL, var.e, var.g, random.seed = NULL)
}
\arguments{
\item{ni}{Number of individuals}

\item{context.weights}{Sampling rates for each context. Will append
a 1 to these rates to ensure each sample has
at least 1 context. Total number of contexts
will be \code{length(context.weights) + 1}.}

\item{var.e}{Residual variance}

\item{var.g}{Within-individual residual variance + covariance}

\item{random.seed}{Seed for reproducibility}
}
\value{
A list with slots \code{X}, \code{K}, and \code{Y} serving as
        input for EMMA R package (design matrix, kinship matrix, and 
        response vector, respectively) and slots \code{X.mc} and \code{Y.mc}
        as input for the mcLMM functions. See \code{mc_mle} or 
        \code{mc_remle} documentation for more details on mcLMM input.
}
\description{
Simulates response Y under the mcLMM model with the given parameters.
Univariate LMM with 012 genotype value as covariate (sampled from 
binomial with 0.5 MAF)
}
