% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_null_data}
\alias{simulate_null_data}
\title{Simulate null test data for mcLMM}
\usage{
simulate_null_data(ni, nc, var.e, var.g, random.seed = NULL)
}
\arguments{
\item{ni}{Number of individuals}

\item{nc}{Number of contexts/tissues}

\item{var.e}{Residual variance}

\item{var.g}{Within-individual residual variance + covariance}

\item{random.seed}{Seed for reproducibility}
}
\value{
A list with slots \code{K}, and \code{Y} serving as
        input for EMMA R package (kinship matrix and 
        response vector, respectively) and slot \code{Y.mc}
        as input for the mcLMM functions. See \code{mc_mle} or 
        \code{mc_remle} documentation for more details on mcLMM input.
}
\description{
Simulates response Y under the mcLMM model with the given parameters.
Samples from MVN with Sigma = K*v_g + I*v_e
}
