% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcLMM.R
\name{meta_tissue}
\alias{meta_tissue}
\title{Ultra-fast meta-tissue algorithm}
\usage{
meta_tissue(
  expr,
  geno,
  covs = NULL,
  heuristic = FALSE,
  newRE = TRUE,
  force.iter = FALSE
)
}
\arguments{
\item{expr}{Matrix with individuals as rows and tissues as columns. Missing
gene expression must be NA}

\item{geno}{Vector of genotypes for each individual}

\item{covs}{Matrix with individuals as rows and covariates as columns.}

\item{heuristic}{Boolean. Uses heuristic for scaling standard errors. 
increases sensitivity at the cost of higher FPR.}

\item{newRE}{Boolean. Use new random effects model to perform meta
analyses as discussed in Sul et al.}

\item{force.iter}{Boolean. If TRUE, force iterative method even when there
is no missing data. This is included for testing purposes
only. The optimal non-iterative method for no missing data
is exact and way faster.}
}
\value{
List of estimated coefficients \code{beta}, coefficient correlation
        \code{corr}, and \code{sigma_g}.
}
\description{
Provides identical results as meta-tissue in linear time with respect to
the number of samples rather than cubic. Slight differences may be 
due to different optimization algorithms of the same likelihood function.
}
